requirements.txt
streamlit>=1.29.0
anthropic>=0.18.0
python-dotenv==1.0.1
pytest==8.0.0
pytest-cov==4.1.0
pytest-asyncio==0.23.0
black==24.2.0
flake8==7.0.0
mypy==1.8.0